requirements.txt
langchain
langchain-community
langchain-core
transformers
accelerate
sentencepiece
torch
python-dotenv
streamlit
pandas
plotly